+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
 ## Unreleased
 
 - Added `wiki page push` command for updating a single MediaWiki page.
+- Added help text when `exec` commands are run without arguments (T294851).
 - Added optional telemetry submission via Wikimedia Event intake.
 - Improved on wiki documentation with auto generated command reference.