+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
 
 - Added `wiki page push` command for updating a single MediaWiki page.
 - Added optional telemetry submission via Wikimedia Event intake.
+- Improved error message when `exec` commands are run without running containers.
 - Improved on wiki documentation with auto generated command reference.
 
 ## v0.8.1